feat: fft.ts
